requirements.txt
# Transport Routing API - Production Dependencies
# 
# PURPOSE: Defines all Python packages required for production deployment
#
# KEY COMPONENTS:
# - AWS CDK libraries for infrastructure as code
# - FastAPI framework for REST API development
# - HTTP client libraries for external API integration
# - Data validation and serialization libraries
# - AWS Lambda runtime and observability tools
#
# WHY USED:
# - Pinned versions ensure reproducible builds
# - Production-ready packages only (no dev tools)
# - AWS-optimized dependencies for Lambda environment
# - Minimal footprint for faster cold starts

aws-cdk-lib==2.115.0
constructs>=10.0.0,<11.0.0
boto3==1.34.34
botocore==1.34.34
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
pydantic==2.5.0
python-dateutil==2.8.2
aws-lambda-powertools==2.29.0
mangum==0.17.0
redis==5.0.1
pytest-xdist==3.5.0
pytest-benchmark==4.0.0